Deep_Sea-NN-main/requirements.txt
torch
torchvision
Pillow
tqdm
